🚚rename : club > club 컴포넌트 폴더 위치 조정
