+++ client/src/components/club/club/_createLocal.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
-import { DIVISIONS_DATA } from '../../../../assets/data/divisions';
+import { DIVISIONS_DATA } from '../../../assets/data/divisions';
 import { CreateCategoryProps } from './_createCategory';
-import { S_Label_mg_top } from '../../../../pages/club/club/EditClub';
-import { S_Select } from '../../../UI/S_Select';
+import { S_Label_mg_top } from '../../../pages/club/club/EditClub';
+import { S_Select } from '../../UI/S_Select';
 
 interface DistrictType {
   code: string;
