+++ client/src/pages/club/club/CreateClub.tsx
@@ -4,9 +4,9 @@ import styled from 'styled-components';
 import { postFetch } from '../../../util/api';
 import getGlobalState from '../../../util/authorization/getGlobalState';
 import LoginChecker from '../../../components/LoginChecker';
-import CreateCategory from '../../../components/club/member/club/_createCategory';
-import CreateLocal from '../../../components/club/member/club/_createLocal';
-import CreateTag from '../../../components/club/member/club/_createTag';
+import CreateCategory from '../../../components/club/club/_createCategory';
+import CreateLocal from '../../../components/club/club/_createLocal';
+import CreateTag from '../../../components/club/club/_createTag';
 import S_Container from '../../../components/UI/S_Container';
 import { S_Input } from '../../../components/UI/S_Input';
 import { S_TextArea } from '../../../components/UI/S_TextArea';
