+++ client/src/pages/club/club/EditClub.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import FormData from 'form-data';
 import { getFetch, patchFetch } from '../../../util/api';
 import getGlobalState from '../../../util/authorization/getGlobalState';
-import CreateLocal from '../../../components/club/member/club/_createLocal';
-import CreateTag from '../../../components/club/member/club/_createTag';
+import CreateLocal from '../../../components/club/club/_createLocal';
+import CreateTag from '../../../components/club/club/_createTag';
 import S_Container from '../../../components/UI/S_Container';
 import { S_Title, S_Label, S_Description } from '../../../components/UI/S_Text';
 import { S_Input } from '../../../components/UI/S_Input';
